Revert "CIV-15610 move CARM non prod tests to master pipeline  (#5046)"
This reverts commit 2aa533b9545b1e1ebd7a895b032f1febbb037a73.
